+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 ### Common changes
 
+- Added support for gain additional friendship by complete a quest
 - Added support for custom background texture and font colors for quest details window in quest log menu
 - Added support for new types of quest reward: `Object` and `Weapon`
 - Fixed some issues with multiplayer
@@ -12,12 +13,14 @@
 
 ### For Content Pack modders
 
+- Added field `FriendshipGain` for define how much friendship points for which NPCs player gains after quest completion
 - Added new fields for quest: `RewardType`, `RewardAmount`, `Texture` and `Colors`
 - Added support for custom quest types in content packs (via `<modUid>/<questTypeName>` in `Type` field)
 - Better error messages
 
 ### For SMAPI modders
 
+- Added new API on `CustomQuest` class for define how much friendship points for which NPCs player gains after quest completion
 - Added new APIs for define quest reward type, custom bg texture nad font colors
 - Deprecated `IStateful.ResetState()` and `IManagedQuestApi.GetById` now raise a compilation error when it is used in code
 - Added active state